Update battleShip.cpp
Hey Johnathan! I love your game, it definitely reminded me of my childhood. I couldn't break it, so that is awesome! I think you thought of any possibility of being able to break it. 

So what I was thinking of adding is a Play again option. (keep in mind i am a terrible coder) so I am just going to add it the way I think it might work. I tried it on my computer and it is not exactly working the way it should, but it would be a nice option.